kev_project2::ept

Struct ExtendedPageTable

pub struct ExtendedPageTable(Box<Inner>);

Second level page table that holds guest-physical to host-physical mapping.

Tuple Fields
0: Box<Inner>
Implementations

impl ExtendedPageTable

pub fn new() -> Self

pub fn pa(&self) -> Pa

pub fn map( &mut self, gpa: Gpa, pg: Page, perm: Permission, ) -> Result<(), EptMappingError>

Map pg into va with permission perm.

pub unsafe fn do_map( &mut self, gpa: Gpa, hpa: Pa, perm: Permission, ) -> Result<(), EptMappingError>

Map hpa into pa with permission perm.

Safety

The Pa must point the physical address of allocated page.

pub fn unmap(&mut self, gpa: Gpa) -> Result<Page, EptMappingError>

Unmap the gpa and returns Page that was mapped to gpa.

pub fn walk(&self, gpa: Gpa) -> Result<&EptPte, EptMappingError>

Walk the extended page table and return corresponding eptpte of the gpa if exist.